backend/requirements.txt
# FastAPI y servidor
fastapi==0.116.1
uvicorn[standard]==0.24.0

# Validación y configuración
pydantic==2.5.0
pydantic-settings==2.1.0

# Procesamiento de archivos
pandas==2.2.3
openpyxl==3.1.2
python-multipart==0.0.6

# Base de datos
sqlalchemy[asyncio]==2.0.23
asyncpg==0.29.0
alembic==1.13.1

# Autenticación y seguridad
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Logging estructurado
structlog==23.2.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.27.0

# PDF Generation (opcional)
# weasyprint==61.2

# Cache
redis==5.0.1
aioredis==2.0.1

# File validation
python-magic==0.4.27

# System monitoring
psutil==5.9.8